Moving helpers into a combined file & module
